Large Refactoring: split classes up into individual files, moved everything into namespace Idology. Underlying code is mostly the same, didn't have to change any tests.
