+++ lib/answer.rb
@@ -1,0 +1,9 @@
+module Idology
+  class Answer
+    attr_accessor :text
+
+    def initialize(answer)
+      self.text = answer
+    end
+  end
+end
