+++ lib/error.rb
@@ -1,0 +1,5 @@
+module Idology
+  class Error < RuntimeError; end
+
+  class ServiceError < Error; end
+end
