+++ lib/idology.rb
@@ -1,4 +1,21 @@
-require File.dirname(__FILE__) + "/../base"
-require File.dirname(__FILE__) + "/base"
-require File.dirname(__FILE__) + "/api_request"
-require File.dirname(__FILE__) + "/api_response"
+require 'net/http'
+require 'net/https'
+require 'logger'
+require File.dirname(__FILE__) + "/access_credentials"
+require File.dirname(__FILE__) + "/answer"
+require File.dirname(__FILE__) + "/error"
+require File.dirname(__FILE__) + "/question"
+require File.dirname(__FILE__) + "/service"
+require File.dirname(__FILE__) + "/subject"
+require File.dirname(__FILE__) + "/request/request"
+require File.dirname(__FILE__) + "/request/challenge_questions_request"
+require File.dirname(__FILE__) + "/request/challenge_verification_request"
+require File.dirname(__FILE__) + "/request/search_request"
+require File.dirname(__FILE__) + "/request/verification_questions_request"
+require File.dirname(__FILE__) + "/request/verification_request"
+require File.dirname(__FILE__) + "/response/response"
+require File.dirname(__FILE__) + "/response/challenge_questions_response"
+require File.dirname(__FILE__) + "/response/challenge_verification_response"
+require File.dirname(__FILE__) + "/response/search_response"
+require File.dirname(__FILE__) + "/response/verification_questions_response"
+require File.dirname(__FILE__) + "/response/verification_response"
