+++ lib/question.rb
@@ -1,0 +1,10 @@
+module Idology
+  class Question
+    attr_accessor :prompt, :type, :candidate_answers, :chosen_answer
+
+    def initialize
+      self.prompt = self.type = ""
+      self.candidate_answers = []
+    end
+  end
+end
